rename: calculateMinSizeBasedOnFigmaWidth -> calcSizeBasedOnFigma
