+++ frontend/src/components/ui/projectList/ProjectListProjectInfo.tsx
@@ -2,9 +2,9 @@ import { AVATAR_STYLE } from 'consts/avatarStyle'
 import { occupationList } from 'consts/occupationList'
 import React, { FC } from 'react'
 import styled from 'styled-components'
-import { calculateMinSizeBasedOnFigmaWidth } from 'utils/calculateMinSizeBasedOnFigmaWidth'
-import { calculateMinSizeBasedOnFigmaHeight } from 'utils/calculateMinSizeBasedOnFigmaWidth'
-import { calculateVhBasedOnFigma } from 'utils/calculateMinSizeBasedOnFigmaWidth'
+import { calculateMinSizeBasedOnFigmaWidth } from 'utils/calculateSizeBasedOnFigma'
+import { calculateMinSizeBasedOnFigmaHeight } from 'utils/calculateSizeBasedOnFigma'
+import { calculateVhBasedOnFigma } from 'utils/calculateSizeBasedOnFigma'
 import { UserAvatarIcon } from '../avatar/UserAvatarIcon'
 import { UserCount } from '../avatar/UserCount'
