+++ frontend/src/styles/theme.ts
@@ -1,4 +1,4 @@
-import { calculateMinSizeBasedOnFigmaWidth } from 'utils/calculateMinSizeBasedOnFigmaWidth'
+import { calculateMinSizeBasedOnFigmaWidth } from 'utils/calculateSizeBasedOnFigma'
 import 'styled-components'
 
 export const theme = {
