+++ frontend/src/utils/calculateSizeBasedOnFigma.ts
